refactor: moving services, gallery and inventory GET routes to a new route with no auth
